extract ConversationTeamInfo
